backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
CouchDB==1.2
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
motor==3.7.1
numpy==2.4.3
packaging==26.0
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
reportlab==4.4.10
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
